Elasticsearch index plugin view is served by a form.
- Using form rather than controller allows other modules to add additional information
  on plugin view form.
