+++ elasticsearch_helper_index_management.routing.yml
@@ -9,8 +9,8 @@ elasticsearch_helper_index_management.index.list:
 elasticsearch_helper_index_management.index.view:
   path: '/admin/config/search/elasticsearch_helper/index/{plugin}'
   defaults:
-    _controller: '\Drupal\elasticsearch_helper_index_management\Controller\IndexController::view'
-    _title_callback: '\Drupal\elasticsearch_helper_index_management\Controller\IndexController::viewTitle'
+    _form: '\Drupal\elasticsearch_helper_index_management\Form\IndexViewForm'
+    _title_callback: '\Drupal\elasticsearch_helper_index_management\Form\IndexViewForm::viewTitle'
   options:
     parameters:
       plugin:
